Classes da página de categoria corrigidas
