+++ src/pages/CadastroCategoria/index.jsx
@@ -1,8 +1,9 @@
+import FormularioCategoria from "../../components/FormularioCategoria";
 
 function CadastroCategoria() {
     return (
         <>
-            <h1>Inserir form de cadastro de categorias aqui</h1>
+            <FormularioCategoria />
         </>
     );
 }
